+++ mleap-core/src/test/scala/ml/combust/mleap/core/regression/DecisionTreeRegressionModelSpec.scala
@@ -22,7 +22,7 @@ class DecisionTreeRegressionModelSpec extends FunSpec {
 
   describe("input/output schema") {
     it("has the right input schema") {
-      assert(regression.inputSchema.fields == Seq(StructField("features", TensorType.Double())))
+      assert(regression.inputSchema.fields == Seq(StructField("features", TensorType.Double(5))))
     }
 
     it("has the right output schema") {
