+++ mleap-runtime/src/test/scala/ml/combust/mleap/runtime/transformer/feature/MaxAbsScalerSpec.scala
@@ -40,8 +40,8 @@ class MaxAbsScalerSpec extends FunSpec{
   describe("input/output schema") {
     it("has the correct inputs and outputs") {
       assert(maxAbsScaler.schema.fields ==
-        Seq(StructField("test_vec", TensorType.Double()),
-          StructField("test_normalized", TensorType.Double())))
+        Seq(StructField("test_vec", TensorType.Double(3)),
+          StructField("test_normalized", TensorType.Double(3))))
     }
   }
 }
