+++ mleap-runtime/src/test/scala/ml/combust/mleap/runtime/transformer/feature/MinMaxScalerSpec.scala
@@ -42,8 +42,8 @@ class MinMaxScalerSpec extends FunSpec{
   describe("input/output schema") {
     it("has the correct inputs and outputs") {
       assert(minMaxScaler.schema.fields ==
-        Seq(StructField("test_vec", TensorType.Double()),
-          StructField("test_normalized", TensorType.Double())))
+        Seq(StructField("test_vec", TensorType.Double(3)),
+          StructField("test_normalized", TensorType.Double(3))))
     }
   }
 }
