+++ mleap-runtime/src/test/scala/ml/combust/mleap/runtime/transformer/feature/StandardScalerSpec.scala
@@ -12,8 +12,8 @@ class StandardScalerSpec extends FunSpec {
       val transformer = StandardScaler(shape = NodeShape.vector(3, 3),
         model = StandardScalerModel(None, Some(Vectors.dense(Array(50.0, 20.0, 30.0)))))
       assert(transformer.schema.fields ==
-        Seq(StructField("input", TensorType.Double()),
-          StructField("output", TensorType.Double())))
+        Seq(StructField("input", TensorType.Double(3)),
+          StructField("output", TensorType.Double(3))))
     }
   }
 }
