+++ mleap-runtime/src/test/scala/ml/combust/mleap/runtime/transformer/regression/LinearRegressionSpec.scala
@@ -40,7 +40,7 @@ class LinearRegressionSpec extends FunSpec {
   describe("input/output schema") {
     it("has the correct inputs and outputs") {
       assert(linearRegression.schema.fields ==
-        Seq(StructField("features", TensorType.Double()),
+        Seq(StructField("features", TensorType.Double(3)),
           StructField("prediction", ScalarType.Double)))
     }
   }
